[YUNIKORN]Refactor persistent_volume e2e test suite structure
