+++ test/e2e/persistent_volume/persistent_volume_suite_test.go
@@ -20,6 +20,7 @@ package persistent_volume
 
 import (
 	"path/filepath"
+	"runtime"
 	"testing"
 
 	"github.com/onsi/ginkgo/v2"
@@ -27,6 +28,9 @@ import (
 	"github.com/onsi/gomega"
 
 	"github.com/apache/yunikorn-k8shim/test/e2e/framework/configmanager"
+	"github.com/apache/yunikorn-k8shim/test/e2e/framework/helpers/common"
+	"github.com/apache/yunikorn-k8shim/test/e2e/framework/helpers/k8s"
+	"github.com/apache/yunikorn-k8shim/test/e2e/framework/helpers/yunikorn"
 )
 
 func init() {
@@ -49,3 +53,25 @@ func TestPersistentVolume(t *testing.T) {
 var Ω = gomega.Ω
 var HaveOccurred = gomega.HaveOccurred
 var dev string
+
+var _ = ginkgo.BeforeSuite(func() {
+	_, filename, _, _ := runtime.Caller(0)
+	suiteName = common.GetSuiteName(filename)
+	// Initializing kubectl client
+	kClient = k8s.KubeCtl{}
+	Ω(kClient.SetClient()).To(gomega.BeNil())
+
+	// Initializing rest client
+	restClient = yunikorn.RClient{}
+	Ω(restClient).NotTo(gomega.BeNil())
+	yunikorn.EnsureYuniKornConfigsPresent()
+})
+
+var _ = ginkgo.AfterSuite(func() {
+	// Clean up
+	ginkgo.By("Deleting PVCs and PVs")
+	err := kClient.DeletePVCs(dev)
+	err2 := kClient.DeletePVs(dev)
+	Ω(err).NotTo(HaveOccurred())
+	Ω(err2).NotTo(HaveOccurred())
+})
